+++ src/main/java/lotto/common/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto2.common;
+package lotto.common;
 
 import java.util.Arrays;
 import java.util.Map;
