+++ src/main/java/lotto/domain/LottoMatchResult.java
@@ -1,6 +1,6 @@
-package lotto2.domain;
+package lotto.domain;
 
-import lotto2.common.LottoRank;
+import lotto.common.LottoRank;
 
 import java.util.Map;
 import java.util.Objects;
