+++ src/main/java/lotto/domain/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto2.domain;
+package lotto.domain;
 
 import java.util.Objects;
