+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,6 +1,6 @@
-package lotto2.domain;
+package lotto.domain;
 
-import lotto2.common.LottoRank;
+import lotto.common.LottoRank;
 
 import java.util.*;
