+++ src/test/java/lotto/domain/LottoTicketMachineTest.java
@@ -1,4 +1,4 @@
-package lotto2.domain;
+package lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
